Fix the render problem of deferred pipeline in editor. (#8373)
* Fix the render problem of deferred pipeline in editor.

* add return
